UPDATE: price decimal to int

+++ products/models.py
@@ -24,7 +24,9 @@ class Product(models.Model):
     )
 
     title = models.CharField(max_length=128, blank=False)
-    price = models.DecimalField(max_digits=10, decimal_places=2, default=0.0)
+    # price = models.DecimalField(max_digits=10, decimal_places=2, default=0.0)
+    price = models.IntegerField(default=0)
+
     adresse = models.CharField(max_length=255, blank=True, null=True)
 
     image = models.URLField(blank=True, null=True)
